refactor: reading the top of a file moves into the parser
The editor reads the use lines straight out of the tokens, because a document
being edited does not parse and what is inside a module is wanted exactly then.
The resolver is about to need the same thing for a different reason: a module
has to be read before whoever imports it, and knowing what to read first cannot
itself require a parse.

So it moves where both can reach it, and the rule about where imports are
written lives in one place rather than two.
